backend/requirements.txt
# --- Backend Server ---
fastapi
uvicorn[standard]

python-dotenv

# --- AI Engine (Gemini) ---
# Used in ai_engine.py for genai.GenerativeModel
google-generativeai

# --- Local RAG (SentenceTransformer) ---
# Used in ai_engine.py for the custom RAG pipeline
sentence-transformers
faiss-cpu
torch
numpy

# --- Optional (recommended for smooth ops) ---
pydantic
requests
